feat: Refactor app and layout components
This change refactors the `app.vue` and `default.vue` components to improve the
organization and separation of concerns. The main changes are:

- Moved the update-related logic and UI from `app.vue` to `default.vue`, as the
  update functionality is more closely tied to the layout than the app itself.
- Removed the unused `useUserStore` and `useUpdater` imports from `app.vue`.
- Removed the `UpdateModal` component from `app.vue`, as it is now included in
  the `default.vue` layout.
